Change logic of all endpoints and test them: - Updated logic for all endpoints, including improvements to the image endpoints. - Corrected the query to find the product by \images.pathname\ instead of \_id\. - Fixed image removal logic to filter images based on \pathname\.  - Improved error handling and response messages.
Note: This implementation is considered final, but further adjustments may be made in the future if needed.

+++ backend/src/controllers/product/deleteProduct.ts
@@ -1,15 +1,32 @@
 import { Request, Response } from "express";
 import Product from "../../models/Product";
+import Section from "../../models/Section";
+import mongoose from "mongoose";
 
 const deleteProduct = async (req: Request, res: Response) => {
   try {
     const { productId } = req.params;
-    const product = await Product.findByIdAndDelete(productId);
 
+    if (!mongoose.Types.ObjectId.isValid(productId)) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Invalid product ID" });
+    }
+
+    // Find and delete the product
+    const product = await Product.findByIdAndDelete(productId);
     if (!product) {
       return res
         .status(404)
         .json({ success: false, message: "Product not found" });
+    }
+
+    const sections = await Section.find({ items: productId });
+    for (const section of sections) {
+      section.items = section.items.filter(
+        (item) => item.toString() !== productId
+      );
+      await section.save(); // Save the updated section
     }
 
     res
